Fixed numeric issues
My attention was drawn to a few issues involving floating point numbers, and integer division. Seems in a few cases, I divided two integers without ensuring that the actual division would be using floating point numbers, when that's what I wanted. Without this change, all those calculations would just drop everything after the decimal, when I want to include that in the subsequent calculations. This was caught by the inspection system. I presume this check is a recent addition.

+++ app/src/main/java/com/noprestige/kanaquiz/logs/LogDetailView.java
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2019 T Duke Perry
+ *    Copyright 2021 T Duke Perry
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -102,7 +102,7 @@ public class LogDetailView extends AppCompatActivity
                 LogDetailItem output = new LogDetailItem(activity);
                 output.setFromRecord(record);
 
-                chartSeries.add(new BarEntry(staticLabels.size(), (record.getCorrectAnswers() * 100) /
+                chartSeries.add(new BarEntry(staticLabels.size(), (float) (record.getCorrectAnswers() * 100) /
                         (record.getCorrectAnswers() + record.getIncorrectAnswers())));
 
                 staticLabels.add(record.getQuestion());
